pyproject.toml
# Project Daedalus - Python Project Configuration
# Created by: orpheus497
# Architecture: Hybrid (Phase 1: FastText+Annoy, Phase 2: LLM+RAG+PEFT)

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daedelus"
version = "0.2.0"
description = "Self-contained terminal assistant with adaptive AI micro-model"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "orpheus497", email = "orpheus497@example.com"}
]
keywords = [
    "terminal",
    "assistant",
    "ai",
    "command-line",
    "shell",
    "autocomplete",
    "machine-learning",
    "privacy",
    "offline"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: BSD",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# ========================================
# ALL DEPENDENCIES
# ========================================
dependencies = [
    # Core Python packages (must be first for build)
    "numpy>=1.24.0,<3.0.0",         # BSD - Array operations

    # Embeddings & Vector Search (Phase 1)
    "fasttext>=0.9.3",              # MIT - Word embeddings (0.9.3+ required for Python 3.14+)
    "annoy>=1.17.3",                # Apache 2.0 - Fast approximate nearest neighbors

    # Terminal Integration
    "prompt-toolkit>=3.0.43",       # BSD 3-Clause - Terminal UI components
    "ptyprocess>=0.7.0",            # ISC - PTY handling for command monitoring

    # CLI & Configuration
    "platformdirs>=4.1.0",          # MIT - User data directories
    "pyyaml>=6.0.1",                # MIT - Configuration management
    "click>=8.1.7",                 # BSD 3-Clause - CLI framework

    # UI & Templates
    "rich>=13.0.0",                 # MIT - Terminal formatting and UI
    "textual>=0.40.0",              # MIT - TUI framework for dashboard
    "jinja2>=3.1.0",                # BSD 3-Clause - Template engine

    # HTTP & Progress
    "requests>=2.31.0",             # Apache 2.0 - HTTP client for downloads
    "tqdm>=4.65.0",                 # MIT/MPL - Progress bars
    "huggingface-hub>=0.19.0",      # Apache 2.0 - Model downloads

    # LLM Inference (Phase 2)
    "llama-cpp-python>=0.2.20",     # MIT - Llama.cpp bindings

    # Vector Database (Phase 2)
    "sqlite-vss>=0.1.0",            # Apache 2.0 - SQLite vector similarity search
    "apsw>=3.40.0",                 # ZLIB - Advanced SQLite wrapper

    # PEFT Fine-Tuning (Phase 2)
    "peft>=0.7.0",                  # Apache 2.0 - Parameter-efficient fine-tuning
    "transformers>=4.36.0",         # Apache 2.0 - Hugging Face transformers
    "accelerate>=0.25.0",           # Apache 2.0 - Training acceleration
    "bitsandbytes>=0.41.0",         # MIT - 8-bit optimizers for quantization
]

# ========================================
# DEVELOPMENT DEPENDENCIES
# ========================================
[project.optional-dependencies]
dev = [
    # Code Quality
    "black>=23.0.0",                # MIT - Code formatter
    "ruff>=0.1.0",                  # MIT - Fast linter
    "mypy>=1.5.0",                  # MIT - Type checker
    "isort>=5.12.0",                # MIT - Import sorter
    "pre-commit>=3.5.0",            # MIT - Pre-commit hooks

    # Profiling & Debugging
    "memory-profiler>=0.61.0",      # BSD - Memory profiling
    "line-profiler>=4.1.0",         # BSD - Line-by-line profiling
    "ipdb>=0.13.13",                # BSD - Debugger

    # Documentation
    "sphinx>=7.0.0",                # BSD - Documentation generator
    "sphinx-rtd-theme>=1.3.0",      # MIT - Read The Docs theme
]

# ========================================
# OPTIONAL DEPENDENCIES (Development extras)
# ========================================
# Note: LLM dependencies are now included by default in main dependencies

# ========================================
# PROJECT ENTRY POINTS
# ========================================
[project.scripts]
daedelus = "daedelus.cli.main:main"
daedelus-daemon = "daedelus.daemon.daemon:main"

# ========================================
# PROJECT URLS
# ========================================
[project.urls]
Homepage = "https://github.com/orpheus497/daedelus"
Documentation = "https://daedelus.readthedocs.io"
Repository = "https://github.com/orpheus497/daedelus"
Issues = "https://github.com/orpheus497/daedelus/issues"
Changelog = "https://github.com/orpheus497/daedelus/blob/main/CHANGELOG.md"

# ========================================
# SETUPTOOLS CONFIGURATION
# ========================================
[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.setuptools.package-data]
daedelus = ["py.typed", "*.yaml", "*.sql"]

# ========================================
# BLACK CODE FORMATTER
# ========================================
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | build
  | dist
)/
'''

# ========================================
# RUFF LINTER
# ========================================
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

# ========================================
# MYPY TYPE CHECKER
# ========================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = true
disallow_untyped_calls = false
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true

# ========================================
# ISORT IMPORT SORTER
# ========================================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
